Adding skips to long tests for maphash
